changed file structure for routing and added gatsby layout plugin
