+++ frontend/src/pages/directory/index.js
@@ -1,0 +1,5 @@
+import React from "react"
+
+const Directory = () => <h1>community Directory</h1>
+
+export default Directory
