+++ frontend/src/pages/submitOrganization/index.js
@@ -1,0 +1,13 @@
+import React from "react"
+import classes from "./submitOrganization.module.css"
+
+const SubmitOrganization= () => {
+  return (
+    <section className={classes.SubmitOrganizationContainer}>
+      <h5>Submit Organization</h5>
+
+    </section>
+  )
+}
+
+export default SubmitOrganization
